feat: removed email+pass, overhauled ui for login
