+++ src/lib/auth.ts
@@ -12,9 +12,6 @@ export const auth = betterAuth({
     provider: "pg",
     schema,
   }),
-  emailAndPassword: {
-    enabled: true,
-  },
   socialProviders: {
     github: {
       clientId: env.GITHUB_CLIENT_ID,
